split service into worker.py
