+++ examples/books_scraper.py
@@ -20,8 +20,8 @@ def parse_books(response: Response, pagination: bool = True):
     for article in articles:
         href = article.h3.a["href"]
         url = urljoin(response.request.url, href)
-        yield Request(url=url, callback=parse_book_details, client=HttpXClient)
-        yield Request(url=url, callback=parse_book_details, client=HttpXClient)
+        yield Request(url=url, callback=parse_book_details, client=HttpXClient())
+        yield Request(url=url, callback=parse_book_details, client=HttpXClient())
     if pagination:
         yield from parse_pagination(response)
 
@@ -30,7 +30,7 @@ def parse_pagination(response):
     next_page = response.soup.find("li", {"class": "next"})
     if next_page is not None:
         next_page_url = urljoin(response.request.url, next_page.a["href"])
-        yield Request(url=next_page_url, callback=parse_books, client=HttpXClient)
+        yield Request(url=next_page_url, callback=parse_books, client=HttpXClient())
 
 
 def parse_book_details(response: Response):
@@ -58,7 +58,7 @@ def main(args):
             Request(
                 url="https://books.toscrape.com/index.html",
                 callback=partial(parse_books, pagination=args.pagination),
-                client=HttpXClient,
+                client=HttpXClient(),
             )
         ]
     )
